+++ jax/_src/core.py
@@ -639,12 +639,9 @@ def escaped_tracer_error(tracer, detail=None):
 
 
 def check_scalar_conversion(arr: Array):
-  if arr.size != 1:
-    raise TypeError("Only length-1 arrays can be converted to Python scalars.")
-  if arr.shape != ():
-    # Added 2023 September 18.
-    warnings.warn("Conversion of an array with ndim > 0 to a scalar is deprecated, "
-                  "and will error in future.", DeprecationWarning, stacklevel=3)
+  if arr.ndim > 0:
+    raise TypeError("Only scalar arrays can be converted to Python scalars; "
+                    f"got {arr.ndim=}")
 
 
 def check_integer_conversion(arr: Array):