+++ CHANGELOG.md
@@ -22,6 +22,9 @@ Remember to align the itemized text with the first line of an item within a list
   * {func}`jax.numpy.linalg.solve` now shows a deprecation warning for batched 1D
     solves with `b.ndim > 1`. In the future these will be treated as batched 2D
     solves.
+  * Conversion of a non-scalar array to a Python scalar now raises an error, regardless
+    of the size of the array. Previously a deprecation warning was raised in the case of
+    non-scalar arrays of size 1. This follows a similar deprecation in NumPy.
 
 ## jaxlib 0.4.25